requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PySerial==3.5
PyUSB==1.2.1
python-dotenv==1.0.0
matplotlib==3.7.2
Pillow==10.0.0
